pyproject.toml
[project]
name = "coreos-newsletter"
version = "0.1.0"
description = "POC: weekly engineering newsletter from GitHub, GitLab, and Jira"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27.0",
  "python-dotenv>=1.0.0",
  "pydantic>=2.0.0",
  "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
llm = ["google-genai>=1.0.0"]

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"
